move theme file to theme folder
